Custom MP #1: Create `measurements/` folder (#3220)
* Move measurement.py file

* Separate measurements in different files.

* Fix circular imports.

* Separate tests

* Delete two empty files

* Fix docstring

* Move MeasurementValue class

* Fix

* Move tests

* Move tests
